minor fixes

+++ node.py
@@ -675,8 +675,7 @@ class ThreadedTCPRequestHandler(SocketServer.BaseRequestHandler):
                     app_log.info("Consensus opinion list:" + str(consensus_opinion_list))
 
                     consensus = most_common(consensus_opinion_list)
-                    consensus_percentage = float(
-                        consensus_opinion_list.count(float(consensus)) / float(len(consensus_opinion_list))) * 100
+                    consensus_percentage = (consensus_opinion_list.count(consensus) / (len(consensus_opinion_list))) * 100
                     app_log.info("Current active connections: " + str(len(active_pool)))
                     app_log.info("Current block consensus: " + str(consensus) + " = " + str(consensus_percentage) + "%")
                     # consensus pool
